+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/AdaptivePerformanceTweaksCore.java
@@ -19,17 +19,29 @@
 
 package de.markusbordihn.adaptiveperformancetweakscore;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.eventbus.api.IEventBus;
+import net.minecraftforge.fml.DistExecutor;
 import net.minecraftforge.fml.IExtensionPoint;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.network.NetworkConstants;
 
+import de.markusbordihn.adaptiveperformancetweakscore.client.TitleScreenHandler;
 import de.markusbordihn.adaptiveperformancetweakscore.debug.DebugManager;
 
 @Mod(Constants.MOD_ID)
 public class AdaptivePerformanceTweaksCore {
 
+  protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
+
   public AdaptivePerformanceTweaksCore() {
+    final IEventBus forgeEventBus = MinecraftForge.EVENT_BUS;
+
     // Make sure the mod being absent on the other network side does not cause the client to display
     // the server as incompatible
     ModLoadingContext.get().registerExtensionPoint(IExtensionPoint.DisplayTest.class,
@@ -38,6 +50,13 @@ public class AdaptivePerformanceTweaksCore {
 
     // Warn if debugging is enabled and automatically disable debug on prod for performance reasons.
     DebugManager.checkForDebugLogging(Constants.LOG_NAME);
+
+    // Time measurement event.
+    DistExecutor.unsafeRunWhenOn(Dist.CLIENT, () -> () -> {
+      log.info("{} Adding time measurement event listener ...", Constants.LOG_PREFIX);
+      forgeEventBus.addListener(TitleScreenHandler::handleInitScreenEvent);
+      forgeEventBus.addListener(TitleScreenHandler::handleScreenOpenEvent);
+    });
   }
 
 }