+++ core/src/main/java/de/markusbordihn/adaptiveperformancetweakscore/message/WarnMessages.java
@@ -53,4 +53,11 @@ public class WarnMessages {
         loggerName, level);
   }
 
+  public static String disabledOptimizationModWarning(String modName) {
+    return String.format(
+        "🛈 There are known issue with the %s mod!"
+            + " For this reasons all optimizations for related %s entities and items are disabled!",
+        modName, modName);
+  }
+
 }