+++ items/src/main/java/de/markusbordihn/adaptiveperformancetweaksitems/entity/ItemEntityManager.java
@@ -46,6 +46,7 @@ import net.minecraftforge.fml.common.Mod.EventBusSubscriber;
 import net.minecraftforge.event.server.ServerAboutToStartEvent;
 
 import de.markusbordihn.adaptiveperformancetweakscore.CoreConstants;
+import de.markusbordihn.adaptiveperformancetweakscore.entity.CoreItemEntityManager;
 import de.markusbordihn.adaptiveperformancetweakscore.message.WarnMessages;
 import de.markusbordihn.adaptiveperformancetweakscore.server.OptimizationEvent;
 import de.markusbordihn.adaptiveperformancetweakscore.server.ServerLoadEvent;
@@ -99,8 +100,8 @@ public class ItemEntityManager {
 
       // Additional checks for conflicting mods.
       if (CoreConstants.GET_IT_TOGETHER_LOADED) {
-        log.warn(() -> WarnMessages.conflictingFeaturesModWarning(CoreConstants.GET_IT_TOGETHER_NAME,
-            "clusters items in a specific radius"));
+        log.warn(() -> WarnMessages.conflictingFeaturesModWarning(
+            CoreConstants.GET_IT_TOGETHER_NAME, "clusters items in a specific radius"));
       }
 
     } else {
@@ -148,18 +149,17 @@ public class ItemEntityManager {
       return;
     }
 
+    // Make sure the Item is relevant for our use case.
+    ItemEntity itemEntity = (ItemEntity) entity;
+    if (!CoreItemEntityManager.isRelevantItemEntity(itemEntity)) {
+      return;
+    }
+
     // All items has the entity minecraft.item, so we are using the translation key
     // to better distinguish the different types of items and minecraft.item as backup.
-    ItemEntity itemEntity = (ItemEntity) entity;
     String itemName = itemEntity.getItem().getItem().getRegistryName().toString();
     if (itemName == null) {
       itemName = itemEntity.getEncodeId();
-    }
-
-    // Ignore dropped air blocks because these are not used at all by the players.
-    // Warning: Removing the air block is a bad idea, because it's used to pre-reserve the space.
-    if (itemName.equals("block.minecraft.air") || itemName.equals("minecraft:air")) {
-      return;
     }
 
     // Get world name and start processing of data
@@ -274,16 +274,17 @@ public class ItemEntityManager {
       return;
     }
 
+    // Make sure the Item is relevant for our use case.
     ItemEntity itemEntity = (ItemEntity) entity;
+    if (!CoreItemEntityManager.isRelevantItemEntity(itemEntity)) {
+      return;
+    }
+
+    // All items has the entity minecraft.item, so we are using the translation key
+    // to better distinguish the different types of items and minecraft.item as backup.
     String itemName = itemEntity.getItem().getItem().getRegistryName().toString();
     if (itemName == null) {
       itemName = itemEntity.getEncodeId();
-    }
-
-    // Ignore dropped air blocks because these are not used at all by the players.
-    // Warning: Removing the air block is a bad idea, because it's used to pre-reserve the space.
-    if (itemName.equals("block.minecraft.air") || itemName.equals("minecraft:air")) {
-      return;
     }
 
     // Get world name and start processing of data